+++ student.rb
@@ -5,7 +5,7 @@ class Student < Person
 
   def initialize(age, classroom, name, parent_permission)
     @classroom = classroom
-    super(age, name, parent_permission)
+    super(age, name, parent_permission: parent_permission)
   end
 
   def play_hooky
