+++ teacher.rb
@@ -3,9 +3,9 @@ require './person'
 class Teacher < Person
   attr_accessor :specialization
 
-  def initialize(age, specialization, name, parent_permission)
+  def initialize(age, specialization, name, parent_permission: true)
     @specialization = specialization
-    super(age, name, parent_permission)
+    super(age, name, parent_permission:)
   end
 
   def can_use_services?
